move back_substitute to graphix.linalg
